+++ torchmetrics/utilities/checks.py
@@ -494,12 +494,15 @@ def _input_format_classification_one_hot(
     return preds.reshape(num_classes, -1), target.reshape(num_classes, -1)
 
 
-def _check_retrieval_functional_inputs(preds: Tensor, target: Tensor) -> None:
+def _check_retrieval_functional_inputs(
+    preds: Tensor, target: Tensor, allow_non_binary_target: bool = False
+) -> Tuple[Tensor, Tensor]:
     """Check ``preds`` and ``target`` tensors are of the same shape and of the correct dtype.
 
     Args:
         preds: either tensor with scores/logits
         target: tensor with ground true labels
+        allow_non_binary_target: whether to allow target to contain non-binary values
 
     Raises:
         ValueError:
@@ -522,16 +525,14 @@ def _check_retrieval_functional_inputs(preds: Tensor, target: Tensor) -> None:
     if not preds.is_floating_point():
         raise ValueError("`preds` must be a tensor of floats")
 
-    if target.max() > 1 or target.min() < 0:
+    if not allow_non_binary_target and target.max() > 1 or target.min() < 0:
         raise ValueError("`target` must contain `binary` values")
 
     return preds.float().flatten(), target.long().flatten()
 
 
 def _check_retrieval_inputs(
-    indexes: Tensor,
-    preds: Tensor,
-    target: Tensor,
+    indexes: Tensor, preds: Tensor, target: Tensor, allow_non_binary_target: bool = False
 ) -> Tuple[Tensor, Tensor, Tensor]:
     """Check ``indexes``, ``preds`` and ``target`` tensors are of the same shape and of the correct dtype.
 
@@ -565,7 +566,7 @@ def _check_retrieval_inputs(
     if target.dtype not in (torch.bool, torch.long, torch.int):
         raise ValueError("`target` must be a tensor of booleans or integers")
 
-    if target.max() > 1 or target.min() < 0:
+    if not allow_non_binary_target and target.max() > 1 or target.min() < 0:
         raise ValueError("`target` must contain `binary` values")
 
     return indexes.long().flatten(), preds.float().flatten(), target.long().flatten()